# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for PathForge - project-based learning app
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LearnerProfile {
  id: number;
  clerkUserId: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  educationLevel?: string | null;
  /** @nullable */
  major?: string | null;
  interests: string;
  experience: string;
  isComplete: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface LearnerProfileInput {
  /** @nullable */
  age?: number | null;
  /** @nullable */
  educationLevel?: string | null;
  /** @nullable */
  major?: string | null;
  interests: string;
  experience: string;
  isComplete: boolean;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  draft: 'draft',
  active: 'active',
  completed: 'completed',
} as const;

export interface Project {
  id: number;
  clerkUserId: string;
  title: string;
  ideaPrompt: string;
  status: ProjectStatus;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  ideaPrompt: string;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  draft: 'draft',
  active: 'active',
  completed: 'completed',
} as const;

export interface ProjectUpdate {
  /** @minLength 1 */
  title?: string;
  status?: ProjectUpdateStatus;
}

export interface ProjectStats {
  projectId: number;
  totalNodes: number;
  completedNodes: number;
  availableNodes: number;
  extraNodes: number;
  progressPercent: number;
}

export type NodeStatus = typeof NodeStatus[keyof typeof NodeStatus];


export const NodeStatus = {
  locked: 'locked',
  available: 'available',
  completed: 'completed',
} as const;

export interface Node {
  id: number;
  mapId: number;
  title: string;
  brief: string;
  status: NodeStatus;
  isExtra: boolean;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  positionX?: number | null;
  /** @nullable */
  positionY?: number | null;
  createdAt: string;
}

export interface NodeEdge {
  id: number;
  fromNodeId: number;
  toNodeId: number;
}

export interface NodeMap {
  projectId: number;
  nodes: Node[];
  edges: NodeEdge[];
}