# lib/api-zod/src/generated/types/learnerProfile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for PathForge - project-based learning app
 * OpenAPI spec version: 0.1.0
 */

export interface LearnerProfile {
  id: number;
  clerkUserId: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  educationLevel?: string | null;
  /** @nullable */
  major?: string | null;
  interests: string;
  experience: string;
  isComplete: boolean;
  createdAt: string;
  updatedAt: string;
}