# lib/api-zod/src/generated/types/node.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for PathForge - project-based learning app
 * OpenAPI spec version: 0.1.0
 */
import type { NodeStatus } from './nodeStatus';

export interface Node {
  id: number;
  mapId: number;
  title: string;
  brief: string;
  status: NodeStatus;
  isExtra: boolean;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  positionX?: number | null;
  /** @nullable */
  positionY?: number | null;
  createdAt: string;
}