# lib/api-zod/src/generated/types/nodeEdge.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for PathForge - project-based learning app
 * OpenAPI spec version: 0.1.0
 */

export interface NodeEdge {
  id: number;
  fromNodeId: number;
  toNodeId: number;
}