# lib/api-zod/src/generated/types/nodeMap.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for PathForge - project-based learning app
 * OpenAPI spec version: 0.1.0
 */
import type { Node } from './node';
import type { NodeEdge } from './nodeEdge';

export interface NodeMap {
  projectId: number;
  nodes: Node[];
  edges: NodeEdge[];
}