# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for PathForge - project-based learning app
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectStats {
  projectId: number;
  totalNodes: number;
  completedNodes: number;
  availableNodes: number;
  extraNodes: number;
  progressPercent: number;
}