# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for PathForge - project-based learning app
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectUpdateStatus } from './projectUpdateStatus';

export interface ProjectUpdate {
  /** @minLength 1 */
  title?: string;
  status?: ProjectUpdateStatus;
}